Implement protocol handler with factory
