+++ src/brain/index_tuner.cpp
@@ -12,7 +12,6 @@
 #include "brain/index_tuner.h"
 
 #include <algorithm>
-#include <unordered_map>
 #include <include/brain/brain_util.h>
 
 #include "concurrency/transaction_manager_factory.h"
@@ -26,7 +25,7 @@
 #include "index/index_factory.h"
 #include "storage/data_table.h"
 #include "storage/tile_group.h"
-#include "network/protocol_handler.h"
+#include "network/postgres_protocol_handler.h"
 
 namespace peloton {
 namespace brain {
@@ -388,8 +387,8 @@ void IndexTuner::AddIndexes(
         // At some point the PreparedStatement handles should be moved out of
         // the ProtocolHandler and into some more sane that doesn't require us
         // to start up the network layer to test...
-        for (auto pm : network::ProtocolHandler::GetPacketManagers()) {
-          pm->InvalidatePreparedStatements(index->GetMetadata()->GetTableOid());
+        for (auto ph : network::PostgresProtocolHandler::GetPostgresProtocolHandlers()) {
+          ph->InvalidatePreparedStatements(index->GetMetadata()->GetTableOid());
         }  // FOR
       }
     }
