+++ test/binder/binder_test.cpp
@@ -20,7 +20,7 @@
 #include "expression/tuple_value_expression.h"
 #include "optimizer/optimizer.h"
 #include "parser/postgresparser.h"
-#include "tcop/tcop.h"
+#include "include/traffic_cop/traffic_cop.h"
 
 #include <sql/testing_sql_util.h>
