+++ test/executor/copy_test.cpp
@@ -23,7 +23,7 @@
 #include "optimizer/rule.h"
 #include "parser/postgresparser.h"
 #include "planner/seq_scan_plan.h"
-#include "tcop/tcop.h"
+#include "include/traffic_cop/traffic_cop.h"
 
 #include "gtest/gtest.h"
 #include "statistics/testing_stats_util.h"
