+++ test/optimizer/old_optimizer_test.cpp
@@ -11,8 +11,8 @@
 #include "parser/postgresparser.h"
 #include "planner/plan_util.h"
 #include "planner/update_plan.h"
-#include "tcop/tcop.h"
 #include <sql/testing_sql_util.h>
+#include "include/traffic_cop/traffic_cop.h"
 
 namespace peloton {
 namespace test {
