+++ test/optimizer/optimizer_test.cpp
@@ -15,8 +15,8 @@
 #include "planner/delete_plan.h"
 #include "planner/insert_plan.h"
 #include "planner/update_plan.h"
-#include "tcop/tcop.h"
 #include <sql/testing_sql_util.h>
+#include "include/traffic_cop/traffic_cop.h"
 
 namespace peloton {
 namespace test {
