+++ test/sql/testing_sql_util.cpp
@@ -19,7 +19,8 @@
 #include "optimizer/rule.h"
 #include "parser/postgresparser.h"
 #include "planner/plan_util.h"
-#include "tcop/tcop.h"
+#include "include/traffic_cop/traffic_cop.h"
+#include "sql/testing_sql_util.h"
 
 namespace peloton {
